+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/CHANGELOG.md
@@ -17,6 +17,9 @@
 
 ### Breaking Changes
 
+* [Request and Dependency Success criteria will now be decided based on
+  `Activity.Status`](https://github.com/Azure/azure-sdk-for-net/pull/31024)
+
 ### Bugs Fixed
 
 * [Remove populating _MS.AggregationIntervalMs to all metrics](https://github.com/Azure/azure-sdk-for-net/pull/29473)